Updated handles_test.go to connect to database and create a member to test that the login works.
