--- db/tests/populate_for_tests.sql
@@ -1,1 +1,0 @@
-INSERT INTO members(name, email, password) VALUES('Test User', 'test@gmail.com', 'supersecret') RETURNING *;
